Fixed a connection issue
This issue is due to the Client -> Server Protocol. The client sends three messages with the opcodes 254, 255 and 80 when it connects. Opcode 255 sends an unsigned integer "initialise key" to the server which is like a handshake, to verify you are a valid user.
The "initialise key" used now is "2200049715" but due to some unknown reason, it doesn't let you connect to the server. I have replaced this with "154669603" which fixes this and lets you to connect to the server.