Wall code re-integration + modification
The blob is always centered in the fake borders (the restricted map).
This is true, except when you are close to a side/corner, when one or of the side will be the real borders. The wall code would work as it would only get activated when we are close to the real border and to make this work, I have modified the wall code so that the walls are detected from a smaller distance.