Virus detection code changed
The number has been brought down to 1.1 for several reasons. If the value is mathematically perfect then the bot can hover above a virus (as it can't eat it, it won't think it's a threat) but often there are people under the virus and it will go for them (remember for the bot the virus doesn't exist). Also even if there is no player hiding under the virus, if it hovers over a virus, people tend to try and feed you just to make you pop.